[Other]Right Rotate a list
This is a python implementation to right rotate a list of n elements by k positions clockwise. The function rightRotate() returns the rotated list. The approach used in here is to traverse the elements in the first list one by one and then put the elements at required places in a second list.