[HackerRank] Simple Array Sum Challenge
Function Description

Completed the simpleArraySum function. It returns the sum of the array elements as an integer.
simpleArraySum has the following parameter(s):

ar: an array of integers

Input Format
The first line contains an integer, , denoting the size of the array.
The second line contains  space-separated integers representing the array's elements.